separated common ajax and admin ajax class
